rename azios.js to api.js
